+++ cpp/include/typedb.hpp
@@ -22,44 +22,44 @@
 
 #pragma once
 
-#include "typedb/common/errormessage.hpp"
+#include "typedb/common/error_message.hpp"
 #include "typedb/common/exception.hpp"
 #include "typedb/common/future.hpp"
 #include "typedb/common/iterator.hpp"
 #include "typedb/common/native.hpp"
 
-#include "typedb/connection/databasemanager.hpp"
+#include "typedb/connection/database_manager.hpp"
 #include "typedb/connection/driver.hpp"
 #include "typedb/connection/options.hpp"
 #include "typedb/connection/session.hpp"
 #include "typedb/connection/transaction.hpp"
 
 #include "typedb/user/user.hpp"
-#include "typedb/user/usermanager.hpp"
+#include "typedb/user/user_manager.hpp"
 
-#include "typedb/query/querymanager.hpp"
+#include "typedb/query/query_manager.hpp"
 
 #include "typedb/logic/explainable.hpp"
 #include "typedb/logic/explanation.hpp"
-#include "typedb/logic/logicmanager.hpp"
+#include "typedb/logic/logic_manager.hpp"
 #include "typedb/logic/rule.hpp"
 
-#include "typedb/answer/conceptmap.hpp"
-#include "typedb/answer/conceptmapgroup.hpp"
+#include "typedb/answer/concept_map.hpp"
+#include "typedb/answer/concept_map_group.hpp"
 #include "typedb/answer/json.hpp"
-#include "typedb/answer/valuefuture.hpp"
-#include "typedb/answer/valuegroup.hpp"
+#include "typedb/answer/value_future.hpp"
+#include "typedb/answer/value_group.hpp"
 
 #include "typedb/concept/concept.hpp"
-#include "typedb/concept/conceptmanager.hpp"
+#include "typedb/concept/concept_manager.hpp"
 #include "typedb/concept/thing/attribute.hpp"
 #include "typedb/concept/thing/entity.hpp"
 #include "typedb/concept/thing/relation.hpp"
 #include "typedb/concept/thing/thing.hpp"
-#include "typedb/concept/type/attributetype.hpp"
-#include "typedb/concept/type/entitytype.hpp"
-#include "typedb/concept/type/relationtype.hpp"
-#include "typedb/concept/type/roletype.hpp"
-#include "typedb/concept/type/thingtype.hpp"
+#include "typedb/concept/type/attribute_type.hpp"
+#include "typedb/concept/type/entity_type.hpp"
+#include "typedb/concept/type/relation_type.hpp"
+#include "typedb/concept/type/role_type.hpp"
+#include "typedb/concept/type/thing_type.hpp"
 #include "typedb/concept/type/type.hpp"
 #include "typedb/concept/value/value.hpp"
