+++ cpp/include/typedb/common/error_message.hpp
