+++ cpp/include/typedb/concept/concept_manager.hpp
@@ -23,9 +23,9 @@
 #include "typedb/common/future.hpp"
 #include "typedb/common/native.hpp"
 
-#include "typedb/concept/type/attributetype.hpp"
-#include "typedb/concept/type/entitytype.hpp"
-#include "typedb/concept/type/relationtype.hpp"
+#include "typedb/concept/type/attribute_type.hpp"
+#include "typedb/concept/type/entity_type.hpp"
+#include "typedb/concept/type/relation_type.hpp"
 
 #include "typedb/concept/thing/attribute.hpp"
 #include "typedb/concept/thing/entity.hpp"
