+++ cpp/include/typedb/concept/type/role_type.hpp
