+++ cpp/include/typedb/concept/type/thing_type.hpp
