+++ cpp/include/typedb/connection/database_manager.hpp
