+++ cpp/include/typedb/user/user_manager.hpp
