+++ cpp/lib/answer/concept_map_group.cpp
@@ -19,10 +19,10 @@
  * under the License.
  */
 
-#include "typedb/answer/conceptmapgroup.hpp"
+#include "typedb/answer/concept_map_group.hpp"
 #include "typedb/common/exception.hpp"
 
-#include "inc/conceptfactory.hpp"
+#include "inc/concept_factory.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
