+++ cpp/lib/concept/concept_factory.cpp
@@ -19,18 +19,18 @@
  * under the License.
  */
 
-#include "inc/conceptfactory.hpp"
+#include "inc/concept_factory.hpp"
 
 #include "typedb/common/exception.hpp"
 #include "typedb/connection/transaction.hpp"
 
 #include "typedb/concept/type/type.hpp"
 
-#include "typedb/concept/type/attributetype.hpp"
-#include "typedb/concept/type/entitytype.hpp"
-#include "typedb/concept/type/relationtype.hpp"
-#include "typedb/concept/type/roletype.hpp"
-#include "typedb/concept/type/thingtype.hpp"
+#include "typedb/concept/type/attribute_type.hpp"
+#include "typedb/concept/type/entity_type.hpp"
+#include "typedb/concept/type/relation_type.hpp"
+#include "typedb/concept/type/role_type.hpp"
+#include "typedb/concept/type/thing_type.hpp"
 
 #include "typedb/concept/thing/attribute.hpp"
 #include "typedb/concept/thing/entity.hpp"
