+++ cpp/lib/concept/future.cpp
@@ -22,8 +22,8 @@
 #include "typedb/common/exception.hpp"
 #include "typedb/common/native.hpp"
 
-#include "inc/conceptfactory.hpp"
-#include "inc/conceptfuture.hpp"
+#include "inc/concept_factory.hpp"
+#include "inc/concept_future.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
