+++ cpp/lib/concept/iterator.cpp
@@ -22,8 +22,8 @@
 #include "typedb/common/exception.hpp"
 #include "typedb/common/native.hpp"
 
-#include "inc/conceptfactory.hpp"
-#include "inc/conceptiterator.hpp"
+#include "inc/concept_factory.hpp"
+#include "inc/concept_iterator.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
