+++ cpp/lib/concept/type/attribute_type.cpp
@@ -20,12 +20,12 @@
  * under the License.
  */
 
-#include "typedb/concept/type/attributetype.hpp"
+#include "typedb/concept/type/attribute_type.hpp"
 #include "typedb/common/exception.hpp"
 #include "typedb/concept/thing/attribute.hpp"
 #include "typedb/connection/transaction.hpp"
 
-#include "inc/conceptfactory.hpp"
+#include "inc/concept_factory.hpp"
 #include "inc/macros.hpp"
 
 namespace TypeDB {
