+++ cpp/lib/user/user.cpp
@@ -20,7 +20,7 @@
  */
 
 #include "typedb/user/user.hpp"
-#include "typedb/user/usermanager.hpp"
+#include "typedb/user/user_manager.hpp"
 
 #include "inc/macros.hpp"
 #include "inc/utils.hpp"
